requirements.txt
Flask
Flask-CORS
pandas
pyarrow
numpy
geopy
scipy
python-dotenv
gunicorn